Add --l2-anytrust option to run in AnyTrust mode
test-node.bash can now be run with --l2-anytrust which runs the l2
nodes in AnyTrust mode. It creates a committee of 2 daserver with
assumed-honest set to 1, requiring successful stores of the batch data
to both of them. Nitro nodes sync from a mirror daserver. All dasevers
use local file storage for the batch data on their own volumes.

BLS keys for the committee are automatically generated and the keyset is
activated automatically on the SequencerInbox contract.