ExpressLaneAuction contract deploy from dev source
This adds a command for deploying the ExpressLaneAuction contract:
docker compose run scripts deploy-express-lane-auction

Since the contracts branch with ExpressLaneAuction is not published to
NPM this commit adds a way for it to be included from the local
workspace into the scripts docker image with the --dev-contracts flag.
The bash and docker logic here needs to be cleaned up so that it will
use the NPM version when not using the --dev-contracts flag.

The commit includes WIP code for initializing the contract.