Set a better default ReserveSubmissionSeconds
ReserveSubmissionSeconds represents a period BEFORE the auction closes
that the reserve price may be updated for the next round. The previous
setting of 45 seconds means that the reserve price could not be updated
at all after the round started, coupled with the auction closing period
of 15 seconds (round duration is 60 seconds).